refactor: 패키지 구조 변경
도메인 하위 패키지 생성
