+++ src/main/kotlin/com/dodo/dodolistserver/epic/entity/Epic.kt
@@ -1,4 +1,4 @@
-package com.dodo.dodolistserver.epic
+package com.dodo.dodolistserver.epic.entity
 
 import lombok.Getter
 import java.time.LocalDateTime
