+++ src/main/kotlin/com/dodo/dodolistserver/epic/repository/EpicRepository.kt
@@ -1,5 +1,6 @@
-package com.dodo.dodolistserver.epic
+package com.dodo.dodolistserver.epic.repository
 
+import com.dodo.dodolistserver.epic.entity.Epic
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
