+++ src/main/kotlin/com/dodo/dodolistserver/project/dto/ProjectMapper.kt
@@ -1,5 +1,6 @@
-package com.dodo.dodolistserver.project
+package com.dodo.dodolistserver.project.dto
 
+import com.dodo.dodolistserver.project.entity.Project
 import org.mapstruct.Mapper
 import org.mapstruct.Mapping
 import org.mapstruct.Mappings
