+++ src/main/kotlin/com/dodo/dodolistserver/project/repository/ProjectRepository.kt
@@ -1,5 +1,6 @@
-package com.dodo.dodolistserver.project
+package com.dodo.dodolistserver.project.repository
 
+import com.dodo.dodolistserver.project.entity.Project
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
