+++ src/main/kotlin/com/dodo/dodolistserver/task/entity/Task.kt
@@ -1,4 +1,4 @@
-package com.dodo.dodolistserver.task
+package com.dodo.dodolistserver.task.entity
 
 import lombok.Getter
 import java.time.LocalDateTime
